move utils/exceptions.py > middleware/exceptions/middleware.py
