--- docs/lib/reference/utils/exception.rst
@@ -1,5 +1,0 @@
-exception
-=========
-
-.. automodule:: starlite.utils.exception
-    :members:
